Refactor mealplan package
